feat: #22 move table context to utils
- Move table context to utils/contexts
