+++ src/utils/contexts/index.ts
@@ -1,0 +1,9 @@
+import { createContext } from "react";
+
+import type { TableContextValue } from "./types";
+
+const TableContext = createContext<TableContextValue>({
+  orders: []
+});
+
+export { TableContext };
